use TBlobBatch for intermediate data
