+++ src/io/io.cc
@@ -5,6 +5,7 @@
 #include <mxnet/io.h>
 #include <dmlc/registry.h>
 #include "./image_augmenter.h"
+#include "./iter_batchloader.h"
 #include "./iter_prefetcher.h"
 
 // Registers
