+++ tests/python/train/test_conv.py
@@ -73,11 +73,13 @@ get_data.GetMNIST_ubyte()
 train_dataiter = mx.io.MNISTIter(
         image="data/train-images-idx3-ubyte",
         label="data/train-labels-idx1-ubyte",
-        batch_size=batch_size, shuffle=True, silent=False, seed=10)
+        input_shape=(784,),
+        batch_size=batch_size, shuffle=True, flat=True, silent=False, seed=10)
 val_dataiter = mx.io.MNISTIter(
         image="data/t10k-images-idx3-ubyte",
         label="data/t10k-labels-idx1-ubyte",
-        batch_size=batch_size, shuffle=True, silent=False)
+        input_shape=(784,),
+        batch_size=batch_size, shuffle=True, flat=True, silent=False)
 
 def test_mnist():
     acc_train = 0.0
